# docs/api.yaml
openapi: 3.0.1
info:
  title: pokerdice API
  description: pokerdice API
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    authCookie:
      type: apiKey
      in: cookie
      name: AuthCookie
  schemas:
    ApiResponse:
      type: object
      required: [data]
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      properties:
        total: { type: integer, minimum: 0, description: Total de elementos (ignora paginação) }
        limit: { type: integer, minimum: 0 }
        skip: { type: integer, minimum: 0 }
        next: { type: string, format: uri, nullable: true }
        prev: { type: string, format: uri, nullable: true }
        traceId: { type: string, description: ID de correlação para debugging/observabilidade }
    Problem:
      type: object
      required: [type]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }

    # USERS
...
    LobbyCreateInputModel:
      type: object
      required: [name, maxPlayers, maxRounds, timeoutSeconds]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        maxPlayers: { type: integer, minimum: 2 }
        maxRounds: { type: integer, minimum: 1 }
        timeoutSeconds: { type: integer, minimum: 5 }
    LobbyStatus:
      type: string
      enum: [WAITING, ONGOING, FINISHED]
    LobbyDetails:
      type: object
      required:
        [lobbyId,name,hostId,maxPlayers,maxRounds,timeoutSeconds,status,createdAt,minPlayers,playerCount]
      properties:
        lobbyId: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        hostId: { type: integer }
        minPlayers: { type: integer }
        maxPlayers: { type: integer }
        maxRounds: { type: integer }
        timeoutSeconds: { type: integer }
        status: { $ref: '#/components/schemas/LobbyStatus' }
        createdAt: { type: integer, description: epoch ms }
        playerCount: { type: integer }
    LobbyPlayer:
      type: object
      properties:
        userId: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        balance: { type: number, format: float }
    LobbyWithPlayers:
      allOf:
        - $ref: '#/components/schemas/LobbyDetails'
        - type: object
          required: [players]
          properties:
            players:
              type: array
              items: { $ref: '#/components/schemas/LobbyPlayer' }
    JoinLobbyResponse:
      type: object
      required: [status, lobbyId]
      properties:
        status: { type: string, enum: [joined, match-started] }
        lobbyId: { type: integer }
        matchId: { type: integer, nullable: true }

    # MATCH / ROUNDS / TURNS / HANDS
...
paths:
...
  /api/lobby:
    post:
      tags: [Lobbies]
      summary: Create lobby
      operationId: createLobby
      security:
        - bearerAuth: []
        - authCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LobbyCreateInputModel' }
            examples:
              create:
                value:
                  name: "Lobby Teste"
                  description: "Lobby para testes"
                  maxPlayers: 2
                  maxRounds: 5
                  timeoutSeconds: 30
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema: { type: string, format: uri }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/LobbyDetails' }
              examples:
                created:
                  value:
                    data:
                      lobbyId: 4
                      name: "Lobby Teste"
                      description: "Lobby para testes"
                      hostId: 1
                      minPlayers: 2
                      maxPlayers: 2
                      maxRounds: 5
                      timeoutSeconds: 30
                      status: "WAITING"
                      createdAt: 1761091487943
                      playerCount: 1
                    meta: null
        '400':
          description: Invalid input
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /api/lobbies/{id}:
    get:
      tags: [Lobbies]
      summary: Get lobby by id (with players)
      operationId: getLobbyById
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/LobbyWithPlayers' }
              examples:
                ok:
                  value:
                    data:
                      lobbyId: 4
                      name: "Lobby Teste"
                      description: "Lobby para testes"
                      hostId: 1
                      minPlayers: 2
                      maxPlayers: 2
                      maxRounds: 5
                      timeoutSeconds: 30
                      status: "ONGOING"
                      createdAt: 1761091487943
                      playerCount: 2
                      players:
                        - userId: 1
                          name: Alice
                          email: alice@example.com
                          balance: 10.0
                        - userId: 3
                          name: Carol
                          email: carol@example.com
                          balance: 4.0
                    meta: null
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/lobbies/{id}/join:
    post:
      tags: [Lobbies]
      summary: Join lobby (may start match)
      operationId: joinLobby
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '201':
          description: Created (joined or match started)
          headers:
            Location:
              description: >
                Se `status=joined`, aponta para /api/lobbies/{id};
                se `status=match-started`, aponta para /api/matches/{matchId}
              schema: { type: string, format: uri }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/JoinLobbyResponse' }
              examples:
                joined:
                  value:
                    data: { status: joined, lobbyId: 4 }
                    meta: null
                matchStarted:
                  value:
                    data: { status: match-started, lobbyId: 4, matchId: 15 }
                    meta: null
        '404':
          description: Lobby not found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: Conflict (full / already started / already in lobby)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '500':
          description: Internal error (e.g., match start error)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/lobbies/{id}/leave:
    post:
      tags: [Lobbies]
      summary: Leave lobby
      operationId: leaveLobby
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '204':
          description: No Content
        '404':
          description: Lobby not found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: User not in lobby
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  ############################################################
  # MATCHES
  ############################################################
...